contactus page added

+++ store/src/app/store-list/store-list.component.ts
@@ -5,6 +5,7 @@ import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-store-list',
+
   templateUrl: './store-list.component.html',
   styleUrls: ['./store-list.component.css'],
 })
@@ -12,14 +13,17 @@ export class StoreListComponent implements OnInit {
   title: String = 'Store List';
   //store is the model class for a store item
   stores: StoreModel[];
-  
+
   //creating service object for calling getstores()
   constructor(private storeService: StoreService, private router: Router) {}
-  
+
   doUpdate(store) {
-    this.storeService.setter(store);
-    console.log('function called');
-    this.router.navigate(['/edit']);
+    if (confirm('Are you sure you want to update this?')) {
+      this.storeService.setter(store);
+      
+      this.router.navigate(['/edit']);
+      
+    }
   }
 
   ngOnInit(): void {
@@ -28,10 +32,13 @@ export class StoreListComponent implements OnInit {
       this.stores = JSON.parse(JSON.stringify(data));
     });
   }
+
   deleteItem(store) {
-    this.storeService.deleteStore(store._id).subscribe((res) => {
-      this.stores.splice(this.stores.indexOf(store), 1);
-      // this.router.navigate(['/'])
-    });
+    if (confirm('Are you sure you want to delete this?')) {
+      this.storeService.deleteStore(store._id).subscribe((res) => {
+        this.stores.splice(this.stores.indexOf(store), 1);
+        alert('Restaurant deleted');
+      });
+    }
   }
 }
